Moved Comparable to CompressedObjectParcelable
